NRF5 - disable high speed timer in deep-sleep mode
The following requirement from sleep specification is not met on NRF5 boards:
- High-speed clocks are turned off in deep-sleep mode.

Stop timer used to us ticker before going to deep-sleep and enable it after wake-up. This operation should result in disabling Timer clock sources in sleep mode.

According to documentanion:
If the system does not require any of the clocks provided by the HFCLK clock controller, the HFCLK
controller may enter a power saving mode automatically and switch off the selected clock source. This
occurs if all peripherals that require either PCLK1M, PCLK16M are appropriately stopped or disabled, and
the CPU is sleeping and thereby no longer requesting HCLK.