NRF51_DK: fix setting of lp ticker interrupt
According to the manual:
If the COUNTER is N, writing N+2 to a CC register is guaranteed to trigger a COMPARE event at N+2.
If the COUNTER is N, writing N or N+1 to a CC register may not trigger a COMPARE event.

In case when COUNTER is N and value to be written to CC register is less than N+2, then write N+2.