NRF51_DK: set_interrupt fix
The COMPARE event occurs when the value in compare register is N and the counter value changes from N-1 to N. Therefore, the minimal safe difference between the compare value to be set and the current counter value is 2 ticks. This guarantees that the compare trigger is properly setup before the compare condition occurs.