Add fix for modes where clk idle state is high for SPI K66F driver.
This is just workaround which should be replaced with better solution.

In case when clock idle high state is selected, the driver does not set the clk to high state after init, but just before first transmission.
This causes problems since clk is down, when cs is asserted during the first transfer. This workaround transmits dummy symbol after inti, so the clk value is correctly set later.